requirements.txt
streamlit
pandas
numpy
yfinance
ta
scikit-learn
python-dotenv

